# Conflicts:
#	.travis.yml
#	CMakeLists.txt
#	src/common/tiglcommonfunctions.cpp
#	src/common/tiglcommonfunctions.h
#	src/fuselage/CCPACSFuselage.cpp
#	src/fuselage/CCPACSFuselage.h
#	src/guide_curves/CCPACSGuideCurve.cpp
#	src/guide_curves/CCPACSGuideCurve.h
#	src/guide_curves/CCPACSGuideCurves.cpp
#	src/wing/CCPACSWing.cpp
#	src/wing/CCPACSWingComponentSegment.cpp
#	src/wing/CCPACSWingProfile.cpp
#	src/wing/CCPACSWingProfilePointList.cpp
#	src/wing/CCPACSWingProfilePointList.h
#	src/wing/CCPACSWingSegment.cpp
#	src/wing/CCPACSWingSegment.h
#	tests/CMakeLists.txt
#	tests/unittests/tiglCommonFunctions.cpp
#	tests/unittests/tiglWingGuideCurves.cpp

+++ src/math/tiglmathfunctions.cpp
@@ -203,7 +203,7 @@ double class_function_deriv(const double& N1, const double& N2, const int& n, co
 double shape_function(const std::vector<double>& B, const double& x)
 {
     double ret = 0.;
-    int order = B.size()-1;
+    int order = static_cast<int>(B.size()) - 1;
     int i = 0;
     for (std::vector<double>::const_iterator bIT = B.begin(); bIT != B.end(); ++bIT, ++i) {
         ret += *bIT * bernstein_poly(i, order, x);
@@ -217,7 +217,7 @@ double shape_function(const std::vector<double>& B, const double& x)
 double shape_function_deriv(const std::vector<double>& B, const int& n, const double& x)
 {
     double ret = 0.;
-    int order = B.size()-1;
+    int order = static_cast<int>(B.size()) - 1;
     int i = 0;
     for (std::vector<double>::const_iterator bIT = B.begin(); bIT != B.end(); ++bIT, ++i) {
         ret += *bIT * tigl::bernstein_poly_deriv(n, i, order, x);
